Animations are now enabled. ImageView added that will be made visible during animation. ImageView holds the animation over MandelbrotView because of compatibility with AnimationDrawable
DrawBitmapTask has replaced DrawAsyncTask which has been removed.
DrawBitmapTask now divides the task into parts, rather than percents

StopButton renamed to closeButton

Save Button now displays a dialog to enter the name of the file.

RecordButton creates and stores the Animation Frames to external storage.
RecordButton shows a dialog to enter the name of the animation.
The start of the recording is staggered by 300 milliseconds using a Handler to avoid a bug with a visible keyboard and progress bar. see MyHandler comments for details.

Added Load Button that opens a dialog that shows the saved animations.

Load Button sets the drawable of an imageView to an AnimationDrawable. This requires the manifest to request largeHeap=true since this object can take a lot of memory.

PlayButton plays the currently loaded animation.

CloseButton recycles the bitmaps from the AnimationDrawable

Once, an animation is loaded, all fields except play and close are enabled until close is pressed in order to free resources.
ProgressStart now takes a int parameter indicating the number of parts the task is being divided to.

Utility Class has been updated with many methods to read and write to files.

MandelbrotView methods hav been simplified and have a new helper methods for added buttons

Model no longer get file location, Utility does that now.
Model has static method to create model in between two model objects.